Extract build_request_messages.rs from turn.rs (parent #680)
Tenth vertical-slice extraction: move the per-iteration prompt builder
out of turn.rs into a new src/agent/loop_run/build_request_messages.rs
sibling module.

Migrated 5 methods (all free fns over &mut Agent / &Agent):

Entry point (pub(super)):
- build_request_messages — composes the complete
  Vec<ConversationMessage> for the next assistant request, threading
  system prompt + mode policy + working memory + retrieval injections
  + focused-edit anchors + conversation history.

Private helpers:
- append_general_request_context_messages — workspace scaffold +
  working memory + case/anti-pattern retrieval + repo context.
- maybe_send_request_context_pack — photon context-pack POST with PAM
  advisory bridge.
- append_common_request_messages — offline / plan alias / scaffold
  recovery / verifier repair / artifact-directed notes.
- append_focused_edit_request_messages — focused-edit anchor + slice
  notes + filtered history.

Visibility promotions in turn.rs (impl Agent methods → pub(super)):
- mode_policy_message
- forced_small_edit_recovery_message
- artifact_directed_recovery_message
- verifier_repair_policy_message
- artifact_directed_policy_violation_message
- workspace_appears_empty
- working_memory_message
- repo_context_message

No facade re-export (DR3-001). Caller updates:
- reply_retry.rs: `agent.build_request_messages(...)` →
  `super::build_request_messages::build_request_messages(agent, ...)`

`cargo fix --tests` pruned 4 newly-unused imports from turn.rs + 1 from
loop_run.rs.

turn.rs: 4,852 -> 4,533 (-319 LOC)
Cumulative from 39,489: -34,956 LOC (-88.5%)

Verification:
- cargo fmt --check
- cargo clippy --lib --all-targets -- -D warnings
- cargo test --lib (3,114 passed)
